feat(ENUMS.ts): add ENUMS file to centralize command constants refactor(commitlint.ts): update import path to use ENUMS for command constants refactor(config.ts): update import path to use ENUMS for command constants refactor(githook.ts): update import path to use ENUMS for command constants fix(prompts.ts): correct conventional commit keywords instruction text
